move indexer builder to it's own package
