+++ crates/sui-bridge-indexer/src/models.rs
@@ -1,11 +1,14 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
+use diesel::data_types::PgTimestamp;
+use diesel::{Identifiable, Insertable, Queryable, Selectable};
+
+use sui_indexer_builder::Task;
+
 use crate::schema::{
     progress_store, sui_error_transactions, sui_progress_store, token_transfer, token_transfer_data,
 };
-use diesel::data_types::PgTimestamp;
-use diesel::{Identifiable, Insertable, Queryable, Selectable};
 
 #[derive(Queryable, Selectable, Insertable, Identifiable, Debug)]
 #[diesel(table_name = progress_store, primary_key(task_name))]
@@ -14,6 +17,18 @@ pub struct ProgressStore {
     pub checkpoint: i64,
     pub target_checkpoint: i64,
     pub timestamp: Option<PgTimestamp>,
+}
+
+impl From<ProgressStore> for Task {
+    fn from(value: ProgressStore) -> Self {
+        Self {
+            task_name: value.task_name,
+            checkpoint: value.checkpoint as u64,
+            target_checkpoint: value.target_checkpoint as u64,
+            // Ok to unwrap, timestamp is defaulted to now() in database
+            timestamp: value.timestamp.expect("Timestamp not set").0 as u64,
+        }
+    }
 }
 
 #[derive(Queryable, Selectable, Insertable, Identifiable, Debug)]
