+++ crates/sui-indexer-builder/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "sui-indexer-builder"
+version = "0.1.0"
+authors = ["Mysten Labs <build@mystenlabs.com>"]
+license = "Apache-2.0"
+publish = false
+edition = "2021"
+
+[dependencies]
+anyhow.workspace = true
+tokio = { workspace = true, features = ["full"] }
+async-trait.workspace = true
+mysten-metrics.workspace = true
+sui-types.workspace = true
+sui-data-ingestion-core.workspace = true
+tracing.workspace = true
+prometheus.workspace = true
+telemetry-subscribers.workspace = true
