+++ crates/sui-indexer-builder/src/lib.rs
@@ -1,7 +1,8 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::models;
+pub mod indexer_builder;
+pub mod sui_datasource;
 
 #[derive(Clone, Debug)]
 pub struct Task {
@@ -9,18 +10,6 @@ pub struct Task {
     pub checkpoint: u64,
     pub target_checkpoint: u64,
     pub timestamp: u64,
-}
-
-impl From<models::ProgressStore> for Task {
-    fn from(value: models::ProgressStore) -> Self {
-        Self {
-            task_name: value.task_name,
-            checkpoint: value.checkpoint as u64,
-            target_checkpoint: value.target_checkpoint as u64,
-            // Ok to unwrap, timestamp is defaulted to now() in database
-            timestamp: value.timestamp.expect("Timestamp not set").0 as u64,
-        }
-    }
 }
 
 pub trait Tasks {
